+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+
+## v0.1.0 - 2016-10-29
 ### Added
 - Reset usage when a new month begins and start a new log folder #16
 - Reset byte buffer when device not found on router #30